+++ packages/headless/src/features/search/search-slice.test.ts
@@ -21,7 +21,6 @@ import {
   executeSearch,
   ExecuteSearchThunkReturn,
   fetchFacetValues,
-  fetchInstantResults,
   fetchMoreResults,
   fetchPage,
 } from './search-actions';
@@ -595,26 +594,5 @@ describe('search-slice', () => {
       expect(finalState.response.searchUid).toEqual(response.searchUid);
       expect(finalState.searchResponseId).not.toEqual('test');
     });
-  });
-
-  it('fetchInstantResults updates the searchUid, but not the searchResponseId', () => {
-    const searchUid = 'someid';
-    const action = fetchInstantResults.fulfilled(
-      {
-        results: [buildMockResult()],
-        searchUid,
-      },
-      'req_id',
-      {
-        id: '',
-        q: '',
-        maxResultsPerQuery: 2,
-        cacheTimeout: 10000,
-      }
-    );
-
-    const finalState = searchReducer(state, action);
-    expect(finalState.response.searchUid).toEqual(searchUid);
-    expect(finalState.searchResponseId).not.toEqual(searchUid);
   });
 });