+++ packages/headless/src/features/search/search-slice.ts
@@ -2,7 +2,6 @@ import {createReducer, PayloadAction} from '@reduxjs/toolkit';
 import {
   executeSearch,
   fetchFacetValues,
-  fetchInstantResults,
   fetchMoreResults,
   fetchPage,
 } from './search-actions';
@@ -90,9 +89,6 @@ export const searchReducer = createReducer(
     builder.addCase(fetchFacetValues.fulfilled, (state, action) => {
       state.response.facets = action.payload.response.facets;
       state.response.searchUid = action.payload.response.searchUid;
-    });
-    builder.addCase(fetchInstantResults.fulfilled, (state, action) => {
-      state.response.searchUid = action.payload.searchUid;
     });
     builder.addCase(executeSearch.pending, handlePendingSearch);
     builder.addCase(fetchMoreResults.pending, handlePendingSearch);