+++ packages/atomic-angular/projects/atomic-angular/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.19](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@2.3.18...@coveo/atomic-angular@2.3.19) (2023-02-02)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+## [2.3.18](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@2.3.17...@coveo/atomic-angular@2.3.18) (2023-02-02)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+## [2.3.17](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@2.3.16...@coveo/atomic-angular@2.3.17) (2023-02-02)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
 ## [2.3.16](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@2.3.15...@coveo/atomic-angular@2.3.16) (2023-02-01)
 
 **Note:** Version bump only for package @coveo/atomic-angular