+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.19.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@2.18.0...@coveo/atomic@2.19.0) (2023-02-02)
+
+### Features
+
+- **atomic:** add prop to remove ipx focus trap ([#2710](https://github.com/coveo/ui-kit/issues/2710)) ([077f8f9](https://github.com/coveo/ui-kit/commit/077f8f91364f3d1bc3f4467cb04a2cedb499c861))
+
+# [2.18.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@2.17.1...@coveo/atomic@2.18.0) (2023-02-02)
+
+### Features
+
+- **atomic:** quickview better accessibility support ([#2700](https://github.com/coveo/ui-kit/issues/2700)) ([8f1034b](https://github.com/coveo/ui-kit/commit/8f1034b9d70f5adf5b38ceeef7be9307b9f12ddf))
+
+## [2.17.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@2.17.0...@coveo/atomic@2.17.1) (2023-02-02)
+
+### Bug Fixes
+
+- **atomic,headless:** log proper search ID with instant results ([#2709](https://github.com/coveo/ui-kit/issues/2709)) ([3fb2015](https://github.com/coveo/ui-kit/commit/3fb20151d9f2c77f8aac7dee8f4a5e3aefc54d73))
+
 # [2.17.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@2.16.0...@coveo/atomic@2.17.0) (2023-02-01)
 
 ### Features